fix: qrtest -> qritem, memberNo #20
