+++ src/pages/Test.js
@@ -7,7 +7,6 @@ import axios from 'axios';
 import { useState } from 'react';
 
 const purchase_data = {
-    memberNo: 'dd7ae501-0bbe-41d0-b4b7-c3cbded39a2a', //테스트맨
     franchiseCode: '10003', //편의점 계열
     franchiseType: '편의점',
     franchiseName: 'GS25 - 동교점',
@@ -18,23 +17,23 @@ const purchase_data = {
 const Test = () => {
     const socket = new SockJS('http://localhost:8091/ws');
     const stompClient = Stomp.over(socket);
-function send_enter() {
-    stompClient.send(
-        '/topic/sellinfo',
-        {},
-        JSON.stringify({ message: 'seller enter' }),
-    );
-}
-function send_information() {
-    stompClient.send(
-        '/topic/sellinfo',
-        {},
-        JSON.stringify({
-            message: 'purchase information',
-            data: purchase_data,
-        }),
-    );
-}
+    function send_enter() {
+        stompClient.send(
+            '/topic/sellinfo',
+            {},
+            JSON.stringify({ message: 'seller enter' }),
+        );
+    }
+    function send_information() {
+        stompClient.send(
+            '/topic/sellinfo',
+            {},
+            JSON.stringify({
+                message: 'purchase information',
+                data: purchase_data,
+            }),
+        );
+    }
     //판매자 사이트 접속으로 침
     stompClient.connect({}, function (frame) {
         console.log('Connected: ' + frame);
